OGM-1496 Extract class to handle index validation
